Dagger issue resolved
