+++ features/base/src/main/java/com/instantappsamples/feature/base/common/CommonActivity.java
@@ -1,16 +1,17 @@
-package com.instantappsamples.feature.base;
+package com.instantappsamples.feature.base.common;
 
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 
 import com.google.android.instantapps.samples.feature.base.R;
+import com.instantappsamples.feature.base.BaseActivity;
 
-public class CommonActivity extends BaseActivity {
+public class CommonActivity extends BaseCommonActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_common);
-        getBaseActivityComponent().inject(this);
+        getActivityComponent().inject(this);
     }
 }
