+++ features/bye/build.gradle
@@ -32,5 +32,7 @@ android {
 }
 
 dependencies {
+    api "com.google.dagger:dagger:2.11"
+    annotationProcessor "com.google.dagger:dagger-compiler:2.11"
     implementation project(':features:base')
 }
