+++ features/bye/src/main/java/com/instantappsamples/feature/bye/BaseByeActivity.java
@@ -3,11 +3,8 @@ package com.instantappsamples.feature.bye;
 import android.os.Bundle;
 import android.os.PersistableBundle;
 import android.support.annotation.Nullable;
-import android.support.v7.app.AppCompatActivity;
 
 import com.instantappsamples.feature.base.BaseActivity;
-import com.instantappsamples.feature.base.BaseActivityComponent;
-import com.instantappsamples.feature.base.DaggerBaseActivityComponent;
 import com.instantappsamples.feature.base.HelloApplication;
 
 /**
